// api/src/declarations/gsp_ledger/gsp_ledger.did.d.ts
export interface Account {
  'owner' : Principal,
  'subaccount' : [] | [Subaccount],
}
// ...
export type Subaccount = Uint8Array | number[];